% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf_cal6.R
\name{bf_cal6}
\alias{bf_cal6}
\title{A brfss full calc function}
\usage{
bf_cal6(b_tab, variable, recode_statement = NULL, by_variable = NULL,
  by_variable_recode = NULL, num_years = 5, years = NULL,
  finwt_use = FALSE, bin_filter = NULL)
}
\arguments{
\item{variable}{Do you love rbrfss? Defaults to TRUE.}

\item{recode_statement}{Do you love rbrfss? Defaults to TRUE.}

\item{by_variable}{Do you love rbrfss? Defaults to TRUE.}

\item{by_variable_recode}{Do you love rbrfss? Defaults to TRUE.}

\item{num_years}{Do you love rbrfss? Defaults to TRUE.}

\item{years}{Do you love rbrfss? Defaults to TRUE.}

\item{finwt_use}{Do you love rbrfss? Defaults to TRUE.}

\item{bin_filter}{Do you love rbrfss? Defaults to TRUE.}

\item{btab}{Do you love rbrfss? Defaults to TRUE.}
}
\description{
This function allows you to create the brfss survey design and the prevalence table
}
\examples{
bf_cal6()
}
\keyword{rbrffs}
