% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brf_bydem3.3x.R
\name{brf_bydem3.3x}
\alias{brf_bydem3.3x}
\title{A brfss bydem3.3x calc function}
\usage{
brf_bydem3.3x(table, variable, recode_statement = NULL,
  by_variable1 = NULL, by_variable1_recode = NULL, bin_filter = NULL,
  finwt_use = FALSE)
}
\arguments{
\item{variable}{Do you love rbrfss? Defaults to TRUE.}

\item{recode_statement}{Do you love rbrfss? Defaults to TRUE.}

\item{by_variable1}{Do you love rbrfss? Defaults to TRUE.}

\item{by_variable1_recode}{Do you love rbrfss? Defaults to TRUE.}

\item{bin_filter}{Do you love rbrfss? Defaults to TRUE.}

\item{finwt_use}{Do you love rbrfss? Defaults to TRUE.}

\item{btab}{Do you love rbrfss? Defaults to TRUE.}
}
\description{
This function allows you to create the brfss survey design and the prevalence table
}
\examples{
brf_bydem3.3x()
}
\keyword{rbrffs}
