% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brf_calc.R
\name{brf_calc}
\alias{brf_calc}
\title{A brfss full calc function}
\usage{
brf_calc(brf_df, variable, recode_statement = NULL, by_variable = NULL,
  by_variable_recode = NULL, num_years = 5, years = NULL,
  finwt_use = FALSE, bin_filter = NULL)
}
\arguments{
\item{brf_df}{A brfss raw data object as a dataframe (entered as a character).}

\item{variable}{The main variable of interest for summary calculations (entered as a character).}

\item{recode_statement}{A character text in CAR recode format for recoding the variable parameter.}

\item{by_variable}{A secondary variable defining a subet for summary calculation (entered as character).}

\item{by_variable_recode}{A character text in CAR recode format for recoding the by_variable parameter.}

\item{num_years}{A integer specifying the years to include in aggregate for creating the survey design.}

\item{years}{The specific years to include from yr column in survey design (entered as combined character vector).}

\item{finwt_use}{If TRUE, uses a combined final weight calculated from aggregating years in weighting process. Defaults to FALSE.}

\item{bin_filter}{A character match for quickly keeping/filtering a single level of the primary variable. Must match recode label.}
}
\description{
This function combines the creation of the the brfss survey design (with year identification), variable recoding, and common summary means calculations.
}
\examples{
brf_calc()
}
\keyword{rbrffs}
