% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brf_srv_by.R
\name{brf_srv_by}
\alias{brf_srv_by}
\title{A brfss brf_srv_by calc function}
\usage{
brf_srv_by(brf_srvyr_design, variable, recode_statement = NULL,
  by_variable = NULL, by_variable_recode = NULL, bin_filter = NULL,
  finwt_use = FALSE)
}
\arguments{
\item{brf_srvyr_design}{A brfss srvyr design object.}

\item{variable}{The main variable of interest for summary calculations (entered as a character).}

\item{recode_statement}{A character text in CAR recode format for recoding the variable parameter.}

\item{by_variable}{A secondary variable defining a subet for summary calculation (entered as character).}

\item{by_variable_recode}{A character text in CAR recode format for recoding the by_variable parameter.}

\item{bin_filter}{A character match for quickly keeping/filtering a single level of the primary variable. Must match recode label.}

\item{finwt_use}{If TRUE, uses a combined final weight calculated from aggregating years in weighting process. Defaults to FALSE.}
}
\description{
This function allows you to create the brfss survey design and the prevalence table
}
\examples{
brf_srv_by()
}
\keyword{rbrffs}
