% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brf_srv_des.R
\name{brf_srv_des}
\alias{brf_srv_des}
\title{A brfss brfs_srv_des survey design function}
\usage{
brf_srv_des(brf_df, variable, by_variable = NULL, num_years = 5,
  years = NULL, finwt_use = FALSE)
}
\arguments{
\item{brf_df}{A brfss raw data object as a dataframe (entered as a character).}

\item{variable}{The main variable of interest for summary calculations (entered as a character).}

\item{by_variable}{A secondary variable defining a subet for summary calculation (entered as character).}

\item{num_years}{A integer specifying the years to include in aggregate for creating the survey design.}

\item{years}{The specific years to include from yr column in survey design (entered as combined character vector).}

\item{finwt_use}{If TRUE, uses a combined final weight calculated from aggregating years in weighting process. Defaults to FALSE.}
}
\description{
This function allows you to create the brfss survey design and the prevalence table
}
\examples{
brfs_srv_des()
}
\keyword{rbrffs}
