% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brfs_des_fun10.R
\name{brfs_des_fun10}
\alias{brfs_des_fun10}
\title{A brfss brfs_des_fun10 calc function}
\usage{
brfs_des_fun10(b_tab, b_var, by_var = NULL, num_years = 5,
  years = NULL, finwt_use = FALSE)
}
\arguments{
\item{b_tab}{Do you love rbrfss? Defaults to TRUE.}

\item{b_var}{Do you love rbrfss? Defaults to TRUE.}

\item{by_var}{Do you love rbrfss? Defaults to TRUE.}

\item{num_years}{Do you love rbrfss? Defaults to TRUE.}

\item{years}{Do you love rbrfss? Defaults to TRUE.}

\item{finwt_use}{Do you love rbrfss? Defaults to TRUE.}
}
\description{
This function allows you to create the brfss survey design and the prevalence table
}
\examples{
brfs_des_fun10()
}
\keyword{rbrffs}
