% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brflovefunc.R
\name{rbrffs}
\alias{rbrffs}
\title{A brfss Function}
\usage{
rbrffs(love = TRUE)
}
\arguments{
\item{love}{Do you love rbrfss? Defaults to TRUE.}
}
\description{
This function allows you to express your love of brfss
}
\examples{
rbrfss()
}
\keyword{rbrffs}
